+++ src/ui/sectors.js
@@ -1,6 +1,7 @@
 import blessed from '@hp4k1h5/blessed'
 
 import { clear } from '../util/clear.js'
+import { spin } from '../util/spin.js'
 
 /** if no symbol is provided, it should stay in sync with a chart */
 export function buildSectors(ws, options, data) {
@@ -8,16 +9,19 @@ export function buildSectors(ws, options, data) {
 
   options.box = ws.grid.set(...options.yxhw, blessed.text, {
     name: 'sectors',
-    label: `[${options.id} sectors]`,
+    label: `[${options.id}  sectors ${
+      options.pollMs ? ' .. polling ' + spin() : ''
+    }]`,
     // inputs
     keys: false,
+    // input is true for focus rotation
     input: true,
     mouse: true,
     scrollable: true,
     // styles
     tags: true,
     style: {
-      focus: { border: { fg: '#ddf' } },
+      focus: { border: { fg: '#fc5' } },
     },
   })