+++ src/util/spin.js
@@ -1,0 +1,8 @@
+export const spin = (function () {
+  let elements = ['▗', '▖', '▘', '▝']
+  const pushift = () => {
+    elements.push(elements.shift())
+    return elements.join('')
+  }
+  return pushift
+})()